Move all sanity queries to studio/api
